Improving validation, creating rest filter and rest listener.

+++ src/main/groovy/sandbox/payroll/external/interfaceAdapter/webservice/springmvc/EmployeeRestController.groovy
@@ -5,9 +5,10 @@ import org.modelmapper.ModelMapper
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
-import sandbox.payroll.ModelSnapshot
 import sandbox.payroll.Employee
 import sandbox.payroll.EmployeeRepository
+import sandbox.payroll.ModelSnapshot
+import sandbox.validationNotification.ApplicationValidationNotifier
 
 @RequestMapping(value = "payroll")
 @RestController
@@ -18,27 +19,49 @@ class EmployeeRestController {
     private ModelMapper modelMapper = ModelMapper.smartNewFor(EmployeeRestController)
 
     @RequestMapping(value = "/employee", method = RequestMethod.POST)
-    Employee newEmployee(@RequestBody Employee newEmployee) {
-        employeeRepository.add(newEmployee)
-        model.save()
-        return newEmployee;
+    ResponseEntity<Employee> newEmployee(@RequestBody Map mandatoryEmployeeData) {
+        RestControllerValidationListener listener = getValidationListener()
+        def newEmployee = Employee.newEmployee(mandatoryEmployeeData)
+        if(listener.successful()){
+            employeeRepository.add(newEmployee)
+            model.save()
+        }
+        return listener.generateResponse(newEmployee);
     }
 
     @RequestMapping(value = "/employee/{employeeId}", method = RequestMethod.PATCH)
-    Employee changeEmployee(@PathVariable Long employeeId, @RequestBody String changedAttributes) {
+    ResponseEntity<Employee> changeEmployee(@PathVariable Long employeeId, @RequestBody String changedAttributes) {
+        RestControllerValidationListener listener = getValidationListener()
         def changedEmployee = getResource(employeeId)
         patchChangedAttributesInto(changedEmployee, changedAttributes)
-        employeeRepository.update(changedEmployee)
-        model.save()
-        return changedEmployee;
+        if(listener.successful()){
+            employeeRepository.update(changedEmployee)
+            model.save()
+        }
+
+        return listener.generateResponse(changedEmployee)
     }
 
     @RequestMapping(value = "/employee/{employeeId}", method = RequestMethod.DELETE)
     ResponseEntity<Employee> deleteEmployee(@PathVariable Long employeeId) {
+        RestControllerValidationListener listener = getValidationListener()
         Employee changedEmployee = getResource(employeeId)
-        employeeRepository.remove(changedEmployee)
-        model.save()
-        return ResponseEntity.ok(changedEmployee);
+        if(listener.successful()) {
+            employeeRepository.remove(changedEmployee)
+            model.save()
+        }
+        return listener.generateResponse(changedEmployee);
+    }
+
+    @RequestMapping(value = "/employee", method = RequestMethod.GET)
+    Collection<Employee> listEmployees() {
+        return employeeRepository
+    }
+
+    private RestControllerValidationListener getValidationListener() {
+        def listener = new RestControllerValidationListener()
+        ApplicationValidationNotifier.addObserver(listener)
+        listener
     }
 
     private Employee getResource(long employeeId) {
@@ -50,11 +73,6 @@ class EmployeeRestController {
     private void patchChangedAttributesInto(Employee changedEmployee, String changedAttributes) {
         JsonNode changedAttributesNode = new ObjectMapper().readTree(changedAttributes)
         modelMapper.map(changedAttributesNode, changedEmployee)
-    }
-
-    @RequestMapping(value = "/employee", method = RequestMethod.GET)
-    Collection<Employee> listEmployees() {
-        return employeeRepository
     }
 
     @ResponseStatus(value=HttpStatus.NOT_FOUND) //It is a lame that spring mvc do not provide such general exception
